requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
plotly_express